requirements.txt
discord.py>=2.3.0
pytest>=7.0
pytest-asyncio>=0.21
